requirements.txt
fastapi[standard]
uvicorn
langchain-community
langgraph
langchain-google-genai
tavily-python
langgraph-checkpoint-sqlite
langchain-tavily
langchain-exa
exa_py
arxiv
semanticscholar
python-dotenv
bs4
gTTS
pydub
ffmpeg-python
kokoro
soundfile
transformers
datasets
torch
librosa
scipy